+++ components/table/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @dictu/table
 
+## 1.3.0
+
+### Minor Changes
+
+- b474a6c: Changes to table component:
+  - Smaller caption font-size
+  - Word wrap set on table header
+
+## 1.2.0
+
+### Minor Changes
+
+- fd31554: - added fix for wrapping of table cell content
+  - added possibility to align content to end of cell
+
 ## 1.1.0
 
 ### Minor Changes